+++ R/RBesT-package.R
@@ -31,9 +31,11 @@
 #'  \tab `stepsize=0.01,` \tab `adapt_delta` defaults to `0.99` whenever\cr
 #'  \tab `max_treedepth=20)` \tab `RBesT.MC.s2z` is `FALSE`\cr
 #' `RBesT.MC.s2z` \tab `TRUE` \tab sum-to-zero parametrization of the group\cr
-#' \tab \tab random effects; `FALSE` recovers the legacy\cr
-#' \tab \tab sampling scheme\cr
-#' `RBesT.MC.ncp` \tab 1 \tab parametrization: 0=CP, 1=NCP, 2=Automatic  \cr
+#' \tab \tab random effects; `FALSE` uses the conventional representation;\cr
+#' \tab \tab see the references\cr
+#' `RBesT.MC.ncp` \tab 3 \tab group-effect parametrization: 0=CP, 1=NCP,\cr
+#' \tab \tab 2=automatic CP/NCP endpoint from quadrature-derived\cr
+#' \tab \tab per-group fractions, 3=automatic partial centering\cr
 #' `RBesT.MC.init` \tab 1 \tab range of initial uniform \eqn{[-1,1]} is the default  \cr
 #' `RBesT.MC.rescale` \tab `TRUE` \tab Automatic rescaling of raw parameters  \cr
 #' `RBesT.verbose` \tab `FALSE` \tab requests outputs to be more verbose\cr
@@ -59,6 +61,8 @@
 #'
 #' @references
 #' \insertRef{rstan}{RBesT}
+#'
+#' \insertRef{pinkney2026}{RBesT}
 #'
 #' @useDynLib RBesT, .registration = TRUE
 #'